requirements.txt
Flask
langchain
langchain-ollama
langchain-google-genai
langgraph
pandas
python-dotenv
